Fri, Aug 16, 2024, 10:23 PM -04:00

+++ week9_pset/finance/app.py
@@ -225,7 +225,7 @@ def sell():
         db.execute("UPDATE users SET cash = cash + ? WHERE id = ?", totalSale, session["user_id"])
 
         # Removing the entry from the stocks table if the user sold all of their shares.
-        db.execute("DELETE FROM stocks WHERE stock_count = 0")
+        db.execute("DELETE FROM stocks WHERE stock_count = 0 AND user_id = ? AND stock_symbol = ?", session["user_id"], symbol)
 
         # Redirecting the user back to the homepage.
         return redirect("/")
